add docs and serde test
