+++ kimchi/src/tests/framework.rs
@@ -88,7 +88,7 @@ impl TestFramework {
                 let b = DensePolynomial::from_coefficients_vec(coeffs);
                 index.srs.commit_non_hiding(&b, None)
             };
-            Challenge { chals, comm }
+            Challenge::new(chals, comm)
         };
 
         // add the proof to the batch
